fix : Ajout de JsDoc dans projects et réarengement du code
