+++ portfolio-backend/src/projects/projects.controller.ts
@@ -7,15 +7,11 @@ import {
   Param,
   Delete,
   Query,
-  ValidationPipe,
-  UsePipes,
 } from '@nestjs/common';
 import { ProjectsService } from './projects.service';
-import { CreateProjectDto } from './dto/create-project.dto';
-import { UpdateProjectDto } from './dto/update-project.dto';
+import { CreateProjectDto, UpdateProjectDto } from './dto/projectRequest.dto';
 
 @Controller('projects')
-@UsePipes(new ValidationPipe({ transform: true }))
 export class ProjectsController {
   constructor(private readonly projectsService: ProjectsService) {}
 
@@ -24,6 +20,11 @@ export class ProjectsController {
     return this.projectsService.create(createProjectDto);
   }
 
+  /**
+   * Récupère tous les projets.
+   * @param featured Indique si seuls les projets mis en avant doivent être récupérés. De type string car reçu en query param donc dans l'url c'est forcément une string
+   * @returns Une liste de projets. (avec en paramètre featuredBool qui convertit featured en booléen ou undefined)
+   */
   @Get()
   findAll(@Query('featured') featured?: string) {
     const featuredBool =
@@ -41,11 +42,6 @@ export class ProjectsController {
     return this.projectsService.findOne(id);
   }
 
-  @Get(':id/stats')
-  getStats(@Param('id') id: string) {
-    return this.projectsService.getProjectStats(id);
-  }
-
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateProjectDto: UpdateProjectDto) {
     return this.projectsService.update(id, updateProjectDto);
@@ -54,5 +50,11 @@ export class ProjectsController {
   @Delete(':id')
   remove(@Param('id') id: string) {
     return this.projectsService.remove(id);
+  }
+
+  //STATS
+  @Get(':id/stats')
+  getStats(@Param('id') id: string) {
+    return this.projectsService.getProjectStats(id);
   }
 }
