+++ portfolio-backend/src/projects/projects.service.ts
@@ -1,12 +1,21 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
-import { CreateProjectDto } from './dto/create-project.dto';
-import { UpdateProjectDto } from './dto/update-project.dto';
+import { CreateProjectDto, UpdateProjectDto } from './dto/projectRequest.dto';
 
 @Injectable()
 export class ProjectsService {
   constructor(private prisma: PrismaService) {}
 
+  /**
+   * Permet de créer un nouveau projet
+   *
+   * On sépare les technologyIds du reste des données du projet pour gérer la relation many-to-many à part, car prisma ne permet pas de gérer ça automatiquement.
+   * @param createProjectDto Les données du projet à créer envoyées par le client filtrer grace à un DTO
+   * @returns {Promise<Project>}
+   *
+   * @remarks
+   * Avec Prisma, pour gérer une relation many-to-many lors de la création d'un enregistrement, il faut utiliser la syntaxe `connect` ou `create` dans l'objet `data`pour pouvoir créer l'id des technologies associées comme on y a pas accès directement dans le DTO. Et meme si cette données n'est pas directemetn dans le model de Project, on en à besoin pour créer un projet avec des technologies donc pour en mettre le front doit les fournir. Et le DTO récupère cette donnée. Qu'il nous donne maintenant.
+   */
   async create(createProjectDto: CreateProjectDto) {
     const { technologyIds, ...projectData } = createProjectDto;
 
@@ -14,10 +23,13 @@ export class ProjectsService {
     const project = await this.prisma.project.create({
       data: {
         ...projectData,
-        // Fournir des arrays vides par défaut pour les champs obligatoires
-        highlights: projectData.highlights || [],
-        challenges: projectData.challenges || [],
-        learnings: projectData.learnings || [],
+        ...(technologyIds && {
+          technologies: {
+            create: technologyIds.map((techId) => ({
+              technologyId: techId,
+            })),
+          },
+        }),
         // TODO: Implémenter les relations technologies plus tard
       },
       include: {
@@ -47,6 +59,14 @@ export class ProjectsService {
     return project;
   }
 
+  /**
+   * Affiche tous les projets. Si `featured` est fourni, filtre les projets en fonction de cette valeur.
+   *
+   * const where = featured !== undefined ? { featured } : {};
+   * fonction pour gérer le cas où featured est undefined (non fourni) ou boolean (true/false) ainsi si on veut tous les projets ou seulement les projets mis en avant on peut le gérer. Pour admin je pourrais afficher tous les projets et pour le front seulement les mis en avant. Et chosir ceux que je veux afficher.
+   * @param featured
+   * @returns {Promise<Project[]>}
+   */
   async findAll(featured?: boolean) {
     const where = featured !== undefined ? { featured } : {};
 
@@ -144,6 +164,13 @@ export class ProjectsService {
     return project;
   }
 
+  /**
+   * Affiche un projet en fonction de son slug.
+   *
+   * A gérer pour remplacer les findOne qui utilise l'id. Et utiliser le slug à la place, pour plus de lisibilité et SEO.
+   * @param slug permet de récupérer un projet via son slug
+   * @returns {Promise<Project>}
+   */
   async findBySlug(slug: string) {
     const project = await this.prisma.project.findUnique({
       where: { slug },
@@ -184,6 +211,14 @@ export class ProjectsService {
     return project;
   }
 
+  /**
+   * Met à jour un projet existant.
+   *
+   * On sépare les technologyIds du reste des données du projet pour gérer la relation many-to-many à part, car prisma ne permet pas de gérer ça automatiquement.
+   * @param id ID du projet à mettre à jour
+   * @param updateProjectDto modèle partiel des données du projet à mettre à jour
+   * @returns {Promise<Project>}
+   */
   async update(id: string, updateProjectDto: UpdateProjectDto) {
     const { technologyIds, ...projectData } = updateProjectDto;
 
@@ -234,6 +269,17 @@ export class ProjectsService {
     });
   }
 
+  /**
+   * Calcule les statistiques d'un projet.
+   *
+   * La méthode getProjectStats() calcule le nombre total de quêtes (zones + floors) d'un projet.
+   * On utilise la méthode reduce pour additionner le nombre de quêtes dans chaque zone et chaque floor.
+   * Pour chaque zone, on additionne la longueur du tableau quests, et pour chaque floor, la longueur de floorQuests.
+   * On calcule aussi le nombre de quêtes terminées (status === 'DONE'), le taux d'avancement, le temps total Pomodoro, et le nombre de zones/floors.
+   *
+   * @param id ID du projet
+   * @returns {Promise<ProjectStats>}
+   */
   async getProjectStats(id: string) {
     const project = await this.findOne(id);
