--- portfolio-backend/src/zones/dto/update-zone.dto.ts
@@ -1,21 +1,0 @@
-import { IsString, IsOptional, IsInt, Min, IsEnum } from 'class-validator';
-import { ZoneStatus } from '@prisma/client';
-
-export class UpdateZoneDto {
-  @IsOptional()
-  @IsString()
-  name?: string;
-
-  @IsOptional()
-  @IsString()
-  description?: string;
-
-  @IsOptional()
-  @IsInt()
-  @Min(1)
-  order?: number;
-
-  @IsOptional()
-  @IsEnum(ZoneStatus)
-  status?: ZoneStatus;
-}
